@@ -664,7 +664,7 @@ impl<'a> URLHandler<'a> {
                 false
             }
             Err(e) => {
-                panic!("Can't create stream {}", e);
+                panic!("Can't create stream {e}");
             }
         }
     }
@@ -1190,7 +1190,7 @@ mod old {
                     false
                 }
                 Err(e) => {
-                    panic!("Error creating stream {:?}", e);
+                    panic!("Error creating stream {e:?}");
                 }
             }
         }
