@@ -125,7 +125,7 @@ fn nss_dir() -> PathBuf {
         }
         dir
     };
-    assert!(dir.is_dir(), "NSS_DIR {:?} doesn't exist", dir);
+    assert!(dir.is_dir(), "NSS_DIR {dir:?} doesn't exist");
     // Note that this returns a relative path because UNC
     // paths on windows cause certain tools to explode.
     dir
